% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTRL.R
\docType{data}
\name{FTRL}
\alias{FTRL}
\title{Creates FTRL proximal model.}
\format{\code{\link{R6Class}} object.}
\usage{
FTRL
}
\description{
Creates 'Follow the Regularized Leader' model. Only logistic regression implemented at the moment.
}
\section{Fields}{

\describe{
\item{\code{verbose}}{\code{logical = TRUE} whether to display training inforamtion}
}}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
ftrl = FTRL$new(alpha = 0.1, beta = 0.5, lambda = 0, l1_ratio = 1, dropout = 0)
ftrl$partial_fit(X, y)
ftrl$predict(X)
ftrl$coef()
}
}

\section{Methods}{

\describe{
  \item{\code{$new(alpha = 0.1, beta = 0.5, lambda = 0, l1_ratio = 1, dropout = 0)}}{Constructor
  for FTRL model. For description of arguments see \bold{Arguments} section.}
  \item{\code{$partial_fit(X, y, ...)}}{fits/updates model given input matrix \code{X} and target vector \code{y}.
  \code{X} shape = (n_samples, n_features)}
  \item{\code{$predict(X, ...)}}{predicts output \code{X}}
  \item{\code{$coef()}}{ return coefficients of the regression model}
  \item{\code{$dump()}}{create dump of the model (actually \code{list} with current model parameters)}
  \item{\code{$load(x)}}{load/initialize model from dump)}
}
}

\section{Arguments}{

\describe{
 \item{ftrl}{\code{FTRL} object}
 \item{X}{Input sparse matrix - native format is \code{Matrix::RsparseMatrix}.
 If \code{X} is in different format, model will try to convert it to \code{RsparseMatrix}
 with \code{as(X, "RsparseMatrix")} call}
 \item{alpha}{learning rate}
 \item{beta}{learning rate which controls decay. Please refer to FTRL paper for details.
 Usually convergense does not heavily depend on this parameter, so default value 0.5 is safe.}
 \item{lambda}{regularization parameter}
 \item{l1_ratio}{controls L1 vs L2 penalty mixing. 1 = Lasso regression, 0 = Ridge regression. Elastic net is in between.}
 \item{dropout}{dropout - percentage of random features to exclude from each sample. Kind of regularization.}
 \item{n_features}{number of features in model (number of columns in expected model matrix) }
}
}
\examples{
library(Matrix)
library(FTRL)
i = sample(1000, 1000 * 100, TRUE)
j = sample(1000, 1000 * 100, TRUE)
y = sample(c(0, 1), 1000, TRUE)
x = sample(c(-1, 1), 1000 * 100, TRUE)
odd = seq(1, 99, 2)
x[i \%in\% which(y == 1) & j \%in\% odd] = 1
m = sparseMatrix(i = i, j = j, x = x, dims = c(1000, 1000), giveCsparse = FALSE)
X = as(m, "RsparseMatrix")
ftrl = FTRL$new(alpha = 0.01, beta = 0.1, lambda = 10, l1_ratio = 1, dropout = 0, n_features = ncol(m))
ftrl$partial_fit(X, y, nthread = 8)
w = ftrl$coef()
head(w)
sum(w != 0)
p = ftrl$predict(m)
}
\keyword{datasets}

