% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftrl.R
\name{FTRLDenseOptimizer}
\alias{FTRLDenseOptimizer}
\title{FTRL Optimizer}
\usage{
FTRLDenseOptimizer(n_weights, alpha = 0.1, beta = 1, lambda1 = 1,
  lambda2 = 1)
}
\arguments{
\item{n_weights}{number of coefficients to fit}

\item{alpha}{just look at the paper}

\item{beta}{look at the paper}

\item{lambda1}{paper}

\item{lambda2}{paper paper paper

Really not for production}
}
\description{
FTRL Optimizer
}
